qa demo2 version

+++ internal/parsedb/dbstruct.go
@@ -14,7 +14,7 @@ type AppResourceUsageMonitor struct {
 	Backgroundcycletime          int64  `json:"backgroundcycletime"`
 	Facetime                     int64  `json:"facetime"`
 	Foregroundbytesread          int64  `json:"foregroundbytesread"`
-	Foregroundbyteswritten       int    `json:"foregroundbyteswritten"`
+	Foregroundbyteswritten       int64  `json:"foregroundbyteswritten"`
 	Foregroundnumreadoperations  int    `json:"foregroundnumreadoperations"`
 	Foregroundnumwriteoperations int    `json:"foregroundnumwriteoperations"`
 	Foregroundnumberofflushes    int    `json:"foregroundnumberofflushes"`
@@ -505,7 +505,7 @@ type JumpList struct {
 	FullPath      string `json:"fullpath"`
 	ApplicationID string `json:"application_id"`
 	ComputerName  string `json:"computer_name"`
-	FileSize      int    `json:"filesize"`
+	FileSize      int64  `json:"filesize"`
 	EntryID       int    `json:"entry_id"`
 	RecordTime    int    `json:"recordtime"`
 	CreateTime    int    `json:"createtime"`
@@ -605,7 +605,7 @@ type Prefetch struct {
 	LastRunTime        int    `json:"lastruntime"`
 	ProcessPath        string `json:"processpath"`
 	RunCount           int    `json:"runcount"`
-	FileSize           int    `json:"filesize"`
+	FileSize           int64  `json:"filesize"`
 	FolderCreatedTime  int    `json:"foldercreatedtime"`
 	FolderModifiedTime int    `json:"foldermodifiedtime"`
 }
@@ -703,7 +703,7 @@ type TaskSchedule struct {
 	Command       string `json:"command"`
 	LastRunTime   int    `json:"lastruntime"`
 	NextRunTime   int    `json:"nextruntime"`
-	StartBoundary int    `json:"startboundary"`
+	StartBoundary int64  `json:"startboundary"`
 	EndBoundary   int    `json:"endboundary"`
 }
